tests/requirements.txt
# Test requirements for EcoGuard Home Assistant integration
pytest>=7.4.0
pytest-homeassistant-custom-component>=0.13.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-mock>=3.11.1
aioresponses>=0.7.4
